requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
sse-starlette>=2.1.0
redis>=5.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
fakeredis>=2.23.0
pytest>=7.4.0
